requirements.txt
keyboard>=0.13.5
pystray>=0.19.0
pillow>=10.0.0
pyinstaller>=6.0.0
